+++ vhost_user_net/Cargo.toml
@@ -10,10 +10,10 @@ epoll = ">=4.0.1"
 libc = "0.2.71"
 log = "0.4.8"
 net_util = { path = "../net_util" }
+option_parser = { path = "../option_parser" }
 vhost_user_backend = { path = "../vhost_user_backend" }
 vhost_rs = { git = "https://github.com/cloud-hypervisor/vhost", branch = "dragonball", package = "vhost", features = ["vhost-user-slave"] }
 virtio-bindings = "0.1.0"
 virtio-devices = { path = "../virtio-devices" }
 vm-memory = "0.2.1"
-vmm = { path = "../vmm" }
 vmm-sys-util = ">=0.3.1"
